requirements.txt
alembic==1.0.10
Click==7.0
Flask==1.0.3
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
psycopg2-binary==2.8.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.5
Werkzeug==0.15.4
